Add a first pass at moving the worktree picker to the title bar
