+++ crates/git_ui/src/git_picker.rs
@@ -14,18 +14,11 @@ use workspace::{ModalView, Workspace, pane};
 
 use crate::branch_picker::{self, BranchList, DeleteBranch, FilterRemotes};
 use crate::stash_picker::{self, DropStashItem, ShowStashItem, StashList};
-use crate::worktree_picker::{
-    self, DeleteWorktree, WorktreeFromDefault, WorktreeFromDefaultOnWindow, WorktreeList,
-};
 
-actions!(
-    git_picker,
-    [ActivateBranchesTab, ActivateWorktreesTab, ActivateStashTab,]
-);
+actions!(git_picker, [ActivateBranchesTab, ActivateStashTab,]);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum GitPickerTab {
-    Worktrees,
     Branches,
     Stash,
 }
@@ -34,7 +27,6 @@ impl Display for GitPickerTab {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let label = match self {
             GitPickerTab::Branches => "Branches",
-            GitPickerTab::Worktrees => "Worktrees",
             GitPickerTab::Stash => "Stash",
         };
         write!(f, "{}", label)
@@ -47,7 +39,6 @@ pub struct GitPicker {
     repository: Option<Entity<Repository>>,
     width: Rems,
     branch_list: Option<Entity<BranchList>>,
-    worktree_list: Option<Entity<WorktreeList>>,
     stash_list: Option<Entity<StashList>>,
     _subscriptions: Vec<Subscription>,
     popover_style: bool,
@@ -80,7 +71,6 @@ impl GitPicker {
             repository,
             width,
             branch_list: None,
-            worktree_list: None,
             stash_list: None,
             _subscriptions: Vec::new(),
             popover_style,
@@ -94,9 +84,6 @@ impl GitPicker {
         match self.tab {
             GitPickerTab::Branches => {
                 self.ensure_branch_list(window, cx);
-            }
-            GitPickerTab::Worktrees => {
-                self.ensure_worktree_list(window, cx);
             }
             GitPickerTab::Stash => {
                 self.ensure_stash_list(window, cx);
@@ -132,34 +119,6 @@ impl GitPicker {
         self.branch_list.clone().unwrap()
     }
 
-    fn ensure_worktree_list(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Entity<WorktreeList> {
-        if self.worktree_list.is_none() {
-            let worktree_list = cx.new(|cx| {
-                worktree_picker::create_embedded(
-                    self.repository.clone(),
-                    self.workspace.clone(),
-                    self.width,
-                    window,
-                    cx,
-                )
-            });
-
-            let subscription = cx.subscribe(&worktree_list, |this, _, _: &DismissEvent, cx| {
-                if this.tab == GitPickerTab::Worktrees {
-                    cx.emit(DismissEvent);
-                }
-            });
-
-            self._subscriptions.push(subscription);
-            self.worktree_list = Some(worktree_list);
-        }
-        self.worktree_list.clone().unwrap()
-    }
-
     fn ensure_stash_list(
         &mut self,
         window: &mut Window,
@@ -190,9 +149,8 @@ impl GitPicker {
 
     fn activate_next_tab(&mut self, window: &mut Window, cx: &mut Context<Self>) {
         self.tab = match self.tab {
-            GitPickerTab::Worktrees => GitPickerTab::Branches,
             GitPickerTab::Branches => GitPickerTab::Stash,
-            GitPickerTab::Stash => GitPickerTab::Worktrees,
+            GitPickerTab::Stash => GitPickerTab::Branches,
         };
         self.ensure_active_picker(window, cx);
         self.focus_active_picker(window, cx);
@@ -201,8 +159,7 @@ impl GitPicker {
 
     fn activate_previous_tab(&mut self, window: &mut Window, cx: &mut Context<Self>) {
         self.tab = match self.tab {
-            GitPickerTab::Worktrees => GitPickerTab::Stash,
-            GitPickerTab::Branches => GitPickerTab::Worktrees,
+            GitPickerTab::Branches => GitPickerTab::Stash,
             GitPickerTab::Stash => GitPickerTab::Branches,
         };
         self.ensure_active_picker(window, cx);
@@ -217,11 +174,6 @@ impl GitPicker {
                     branch_list.focus_handle(cx).focus(window, cx);
                 }
             }
-            GitPickerTab::Worktrees => {
-                if let Some(worktree_list) = &self.worktree_list {
-                    worktree_list.focus_handle(cx).focus(window, cx);
-                }
-            }
             GitPickerTab::Stash => {
                 if let Some(stash_list) = &self.stash_list {
                     stash_list.focus_handle(cx).focus(window, cx);
@@ -233,30 +185,12 @@ impl GitPicker {
     fn render_tab_bar(&self, cx: &mut Context<Self>) -> impl IntoElement {
         let focus_handle = self.focus_handle(cx);
         let branches_focus_handle = focus_handle.clone();
-        let worktrees_focus_handle = focus_handle.clone();
         let stash_focus_handle = focus_handle;
 
         h_flex().p_2().pb_0p5().w_full().child(
             ToggleButtonGroup::single_row(
                 "git-picker-tabs",
                 [
-                    ToggleButtonSimple::new(
-                        GitPickerTab::Worktrees.to_string(),
-                        cx.listener(|this, _, window, cx| {
-                            this.tab = GitPickerTab::Worktrees;
-                            this.ensure_active_picker(window, cx);
-                            this.focus_active_picker(window, cx);
-                            cx.notify();
-                        }),
-                    )
-                    .tooltip(move |_, cx| {
-                        Tooltip::for_action_in(
-                            "Toggle Worktree Picker",
-                            &ActivateWorktreesTab,
-                            &worktrees_focus_handle,
-                            cx,
-                        )
-                    }),
                     ToggleButtonSimple::new(
                         GitPickerTab::Branches.to_string(),
                         cx.listener(|this, _, window, cx| {
@@ -297,9 +231,8 @@ impl GitPicker {
             .style(ToggleButtonGroupStyle::Outlined)
             .auto_width()
             .selected_index(match self.tab {
-                GitPickerTab::Worktrees => 0,
-                GitPickerTab::Branches => 1,
-                GitPickerTab::Stash => 2,
+                GitPickerTab::Branches => 0,
+                GitPickerTab::Stash => 1,
             }),
         )
     }
@@ -313,10 +246,6 @@ impl GitPicker {
             GitPickerTab::Branches => {
                 let branch_list = self.ensure_branch_list(window, cx);
                 branch_list.into_any_element()
-            }
-            GitPickerTab::Worktrees => {
-                let worktree_list = self.ensure_worktree_list(window, cx);
-                worktree_list.into_any_element()
             }
             GitPickerTab::Stash => {
                 let stash_list = self.ensure_stash_list(window, cx);
@@ -335,13 +264,6 @@ impl GitPicker {
             GitPickerTab::Branches => {
                 if let Some(branch_list) = &self.branch_list {
                     branch_list.update(cx, |list, cx| {
-                        list.handle_modifiers_changed(ev, window, cx);
-                    });
-                }
-            }
-            GitPickerTab::Worktrees => {
-                if let Some(worktree_list) = &self.worktree_list {
-                    worktree_list.update(cx, |list, cx| {
                         list.handle_modifiers_changed(ev, window, cx);
                     });
                 }
@@ -378,45 +300,6 @@ impl GitPicker {
         if let Some(branch_list) = &self.branch_list {
             branch_list.update(cx, |list, cx| {
                 list.handle_filter(&FilterRemotes, window, cx);
-            });
-        }
-    }
-
-    fn handle_worktree_from_default(
-        &mut self,
-        _: &WorktreeFromDefault,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if let Some(worktree_list) = &self.worktree_list {
-            worktree_list.update(cx, |list, cx| {
-                list.handle_new_worktree(false, window, cx);
-            });
-        }
-    }
-
-    fn handle_worktree_from_default_on_window(
-        &mut self,
-        _: &WorktreeFromDefaultOnWindow,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if let Some(worktree_list) = &self.worktree_list {
-            worktree_list.update(cx, |list, cx| {
-                list.handle_new_worktree(true, window, cx);
-            });
-        }
-    }
-
-    fn handle_worktree_delete(
-        &mut self,
-        _: &DeleteWorktree,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if let Some(worktree_list) = &self.worktree_list {
-            worktree_list.update(cx, |list, cx| {
-                list.handle_delete(&DeleteWorktree, window, cx);
             });
         }
     }
@@ -459,11 +342,6 @@ impl Focusable for GitPicker {
                     return branch_list.focus_handle(cx);
                 }
             }
-            GitPickerTab::Worktrees => {
-                if let Some(worktree_list) = &self.worktree_list {
-                    return worktree_list.focus_handle(cx);
-                }
-            }
             GitPickerTab::Stash => {
                 if let Some(stash_list) = &self.stash_list {
                     return stash_list.focus_handle(cx);
@@ -492,7 +370,6 @@ impl Render for GitPicker {
                 key_context.add("GitPicker");
                 match self.tab {
                     GitPickerTab::Branches => key_context.add("GitBranchSelector"),
-                    GitPickerTab::Worktrees => key_context.add("GitWorktreeSelector"),
                     GitPickerTab::Stash => key_context.add("StashList"),
                 }
                 key_context
@@ -517,12 +394,6 @@ impl Render for GitPicker {
                 this.focus_active_picker(window, cx);
                 cx.notify();
             }))
-            .on_action(cx.listener(|this, _: &ActivateWorktreesTab, window, cx| {
-                this.tab = GitPickerTab::Worktrees;
-                this.ensure_active_picker(window, cx);
-                this.focus_active_picker(window, cx);
-                cx.notify();
-            }))
             .on_action(cx.listener(|this, _: &ActivateStashTab, window, cx| {
                 this.tab = GitPickerTab::Stash;
                 this.ensure_active_picker(window, cx);
@@ -533,11 +404,6 @@ impl Render for GitPicker {
             .when(self.tab == GitPickerTab::Branches, |el| {
                 el.on_action(cx.listener(Self::handle_delete_branch))
                     .on_action(cx.listener(Self::handle_filter_remotes))
-            })
-            .when(self.tab == GitPickerTab::Worktrees, |el| {
-                el.on_action(cx.listener(Self::handle_worktree_from_default))
-                    .on_action(cx.listener(Self::handle_worktree_from_default_on_window))
-                    .on_action(cx.listener(Self::handle_worktree_delete))
             })
             .when(self.tab == GitPickerTab::Stash, |el| {
                 el.on_action(cx.listener(Self::handle_drop_stash))
@@ -555,15 +421,6 @@ pub fn open_branches(
     cx: &mut Context<Workspace>,
 ) {
     open_with_tab(workspace, GitPickerTab::Branches, window, cx);
-}
-
-pub fn open_worktrees(
-    workspace: &mut Workspace,
-    _: &zed_actions::git::Worktree,
-    window: &mut Window,
-    cx: &mut Context<Workspace>,
-) {
-    open_with_tab(workspace, GitPickerTab::Worktrees, window, cx);
 }
 
 pub fn open_stash(
@@ -617,9 +474,6 @@ pub fn register(workspace: &mut Workspace) {
             open_with_tab(workspace, GitPickerTab::Branches, window, cx);
         },
     );
-    workspace.register_action(|workspace, _: &zed_actions::git::Worktree, window, cx| {
-        open_with_tab(workspace, GitPickerTab::Worktrees, window, cx);
-    });
     workspace.register_action(|workspace, _: &zed_actions::git::ViewStash, window, cx| {
         open_with_tab(workspace, GitPickerTab::Stash, window, cx);
     });
