+++ crates/git_ui/src/worktree_picker.rs
@@ -1,162 +1,170 @@
+use std::path::PathBuf;
+use std::sync::Arc;
+
 use anyhow::Context as _;
 use collections::HashSet;
 use fuzzy::StringMatchCandidate;
-
 use git::repository::Worktree as GitWorktree;
 use gpui::{
-    Action, App, AsyncWindowContext, Context, DismissEvent, Entity, EventEmitter, FocusHandle,
-    Focusable, InteractiveElement, IntoElement, Modifiers, ModifiersChangedEvent, ParentElement,
-    Render, SharedString, Styled, Subscription, Task, WeakEntity, Window, actions, rems,
+    Action, AnyElement, App, Context, DismissEvent, Entity, EventEmitter, FocusHandle, Focusable,
+    IntoElement, ParentElement, Render, SharedString, Styled, Subscription, Task, Window, actions,
+    rems,
 };
 use picker::{Picker, PickerDelegate, PickerEditorPosition};
-use project::project_settings::ProjectSettings;
-use project::{
-    git_store::{Repository, RepositoryEvent},
-    trusted_worktrees::{PathTrust, TrustedWorktrees},
+use project::Project;
+use project::git_store::RepositoryEvent;
+use ui::{
+    Button, Divider, HighlightedLabel, IconButton, KeyBinding, ListItem, ListItemSpacing, Tooltip,
+    prelude::*,
 };
-use remote::{RemoteConnectionOptions, remote_client::ConnectionIdentifier};
-use remote_connection::{RemoteConnectionModal, connect};
-use settings::Settings;
-use std::{path::PathBuf, sync::Arc};
-use ui::{HighlightedLabel, KeyBinding, ListItem, ListItemSpacing, Tooltip, prelude::*};
-use util::{ResultExt, debug_panic, paths::PathExt};
-use workspace::{
-    ModalView, MultiWorkspace, OpenMode, Workspace, notifications::DetachAndPromptErr,
+use util::ResultExt as _;
+use util::paths::PathExt;
+use workspace::{ModalView, MultiWorkspace, Workspace, notifications::DetachAndPromptErr};
+use zed_actions::{
+    CreateWorktree, NewWorktreeBranchTarget, OpenWorktreeInNewWindow, SwitchWorktree,
 };
 
-use crate::git_panel::show_error_toast;
+actions!(worktree_picker, [DeleteWorktree]);
 
-actions!(
-    git,
-    [
-        WorktreeFromDefault,
-        WorktreeFromDefaultOnWindow,
-        DeleteWorktree
-    ]
-);
-
-pub fn open(
-    workspace: &mut Workspace,
-    _: &zed_actions::git::Worktree,
-    window: &mut Window,
-    cx: &mut Context<Workspace>,
-) {
-    let repository = workspace.project().read(cx).active_repository(cx);
-    let workspace_handle = workspace.weak_handle();
-    workspace.toggle_modal(window, cx, |window, cx| {
-        WorktreeList::new(repository, workspace_handle, rems(34.), window, cx)
-    })
-}
-
-pub fn create_embedded(
-    repository: Option<Entity<Repository>>,
-    workspace: WeakEntity<Workspace>,
-    width: Rems,
-    window: &mut Window,
-    cx: &mut Context<WorktreeList>,
-) -> WorktreeList {
-    WorktreeList::new_embedded(repository, workspace, width, window, cx)
-}
-
-pub struct WorktreeList {
-    width: Rems,
-    pub picker: Entity<Picker<WorktreeListDelegate>>,
-    picker_focus_handle: FocusHandle,
+pub struct WorktreePicker {
+    picker: Entity<Picker<WorktreePickerDelegate>>,
+    focus_handle: FocusHandle,
     _subscriptions: Vec<Subscription>,
-    embedded: bool,
 }
 
-impl WorktreeList {
-    fn new(
-        repository: Option<Entity<Repository>>,
-        workspace: WeakEntity<Workspace>,
-        width: Rems,
+impl WorktreePicker {
+    pub fn new(project: Entity<Project>, window: &mut Window, cx: &mut Context<Self>) -> Self {
+        Self::new_inner(project, false, window, cx)
+    }
+
+    pub fn new_modal(
+        project: Entity<Project>,
         window: &mut Window,
         cx: &mut Context<Self>,
     ) -> Self {
-        let mut this = Self::new_inner(repository, workspace, width, false, window, cx);
-        this._subscriptions
-            .push(cx.subscribe(&this.picker, |_, _, _, cx| {
-                cx.emit(DismissEvent);
-            }));
-        this
+        Self::new_inner(project, true, window, cx)
     }
 
     fn new_inner(
-        repository: Option<Entity<Repository>>,
-        workspace: WeakEntity<Workspace>,
-        width: Rems,
-        embedded: bool,
+        project: Entity<Project>,
+        show_footer: bool,
         window: &mut Window,
         cx: &mut Context<Self>,
     ) -> Self {
+        let project_worktree_paths: HashSet<PathBuf> = project
+            .read(cx)
+            .visible_worktrees(cx)
+            .map(|wt| wt.read(cx).abs_path().to_path_buf())
+            .collect();
+
+        let has_multiple_repositories = project.read(cx).repositories(cx).len() > 1;
+
+        let current_branch_name = project.read(cx).active_repository(cx).and_then(|repo| {
+            repo.read(cx)
+                .branch
+                .as_ref()
+                .map(|branch| branch.name().to_string())
+        });
+
+        let repository = if has_multiple_repositories {
+            None
+        } else {
+            project.read(cx).active_repository(cx)
+        };
+
         let all_worktrees_request = repository
             .clone()
-            .map(|repository| repository.update(cx, |repository, _| repository.worktrees()));
+            .map(|repo| repo.update(cx, |repo, _| repo.worktrees()));
 
-        let default_branch_request = repository.clone().map(|repository| {
-            repository.update(cx, |repository, _| repository.default_branch(false))
-        });
+        let default_branch_request = repository
+            .clone()
+            .map(|repo| repo.update(cx, |repo, _| repo.default_branch(false)));
 
-        cx.spawn_in(window, async move |this, cx| {
-            let all_worktrees: Vec<_> = all_worktrees_request
-                .context("No active repository")?
-                .await??
-                .into_iter()
-                .filter(|worktree| !worktree.is_bare) // hide bare repositories
-                .collect();
+        let initial_matches = vec![WorktreeEntry::CreateFromCurrentBranch];
 
-            let default_branch = default_branch_request
-                .context("No active repository")?
-                .await
-                .map(Result::ok)
-                .ok()
-                .flatten()
-                .flatten();
+        let delegate = WorktreePickerDelegate {
+            matches: initial_matches,
+            all_worktrees: Vec::new(),
+            project_worktree_paths,
+            selected_index: 0,
+            project,
+            current_branch_name,
+            default_branch_name: None,
+            has_multiple_repositories,
+            focus_handle: cx.focus_handle(),
+            show_footer,
+        };
 
-            this.update_in(cx, |this, window, cx| {
-                this.picker.update(cx, |picker, cx| {
-                    picker.delegate.all_worktrees = Some(all_worktrees);
-                    picker.delegate.default_branch = default_branch;
-                    picker.delegate.refresh_forbidden_deletion_path(cx);
-                    picker.refresh(window, cx);
-                })
-            })?;
-
-            anyhow::Ok(())
-        })
-        .detach_and_log_err(cx);
-
-        let delegate = WorktreeListDelegate::new(workspace, repository.clone(), window, cx);
         let picker = cx.new(|cx| {
-            Picker::uniform_list(delegate, window, cx)
-                .show_scrollbar(true)
-                .modal(!embedded)
+            Picker::list(delegate, window, cx)
+                .list_measure_all()
+                .modal(false)
+                .max_height(Some(rems(20.).into()))
         });
+
         let picker_focus_handle = picker.focus_handle(cx);
         picker.update(cx, |picker, _| {
-            picker.delegate.focus_handle = picker_focus_handle.clone();
+            picker.delegate.focus_handle = picker_focus_handle;
         });
 
         let mut subscriptions = Vec::new();
+
+        {
+            let picker_handle = picker.downgrade();
+            cx.spawn_in(window, async move |_this, cx| {
+                let all_worktrees: Vec<_> = match all_worktrees_request {
+                    Some(req) => match req.await {
+                        Ok(Ok(worktrees)) => {
+                            worktrees.into_iter().filter(|wt| !wt.is_bare).collect()
+                        }
+                        Ok(Err(err)) => {
+                            log::warn!("WorktreePicker: git worktree list failed: {err}");
+                            return anyhow::Ok(());
+                        }
+                        Err(_) => {
+                            log::warn!("WorktreePicker: worktree request was cancelled");
+                            return anyhow::Ok(());
+                        }
+                    },
+                    None => Vec::new(),
+                };
+
+                let default_branch = match default_branch_request {
+                    Some(req) => req.await.ok().and_then(Result::ok).flatten(),
+                    None => None,
+                };
+
+                picker_handle.update_in(cx, |picker, window, cx| {
+                    picker.delegate.all_worktrees = all_worktrees;
+                    picker.delegate.default_branch_name =
+                        default_branch.map(|branch| branch.to_string());
+                    picker.refresh(window, cx);
+                })?;
+
+                anyhow::Ok(())
+            })
+            .detach_and_log_err(cx);
+        }
+
         if let Some(repo) = &repository {
-            let picker_entity = picker.clone();
-            subscriptions.push(cx.subscribe(
+            let picker_entity = picker.downgrade();
+            subscriptions.push(cx.subscribe_in(
                 repo,
-                move |_this, repo, event: &RepositoryEvent, cx| {
+                window,
+                move |_this, repo, event: &RepositoryEvent, window, cx| {
                     if matches!(event, RepositoryEvent::GitWorktreeListChanged) {
                         let worktrees_request = repo.update(cx, |repo, _| repo.worktrees());
                         let picker = picker_entity.clone();
-                        cx.spawn(async move |_, cx| {
+                        cx.spawn_in(window, async move |_, cx| {
                             let all_worktrees: Vec<_> = worktrees_request
                                 .await??
                                 .into_iter()
-                                .filter(|worktree| !worktree.is_bare)
+                                .filter(|wt| !wt.is_bare)
                                 .collect();
-                            picker.update(cx, |picker, cx| {
-                                picker.delegate.all_worktrees = Some(all_worktrees);
-                                picker.delegate.refresh_forbidden_deletion_path(cx);
-                            });
+                            picker.update_in(cx, |picker, window, cx| {
+                                picker.delegate.all_worktrees = all_worktrees;
+                                picker.refresh(window, cx);
+                            })?;
                             anyhow::Ok(())
                         })
                         .detach_and_log_err(cx);
@@ -165,386 +173,141 @@ impl WorktreeList {
             ));
         }
 
+        subscriptions.push(cx.subscribe(&picker, |_, _, _, cx| {
+            cx.emit(DismissEvent);
+        }));
+
         Self {
+            focus_handle: picker.focus_handle(cx),
             picker,
-            picker_focus_handle,
-            width,
             _subscriptions: subscriptions,
-            embedded,
         }
     }
-
-    fn new_embedded(
-        repository: Option<Entity<Repository>>,
-        workspace: WeakEntity<Workspace>,
-        width: Rems,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Self {
-        let mut this = Self::new_inner(repository, workspace, width, true, window, cx);
-        this._subscriptions
-            .push(cx.subscribe(&this.picker, |_, _, _, cx| {
-                cx.emit(DismissEvent);
-            }));
-        this
-    }
-
-    pub fn handle_modifiers_changed(
-        &mut self,
-        ev: &ModifiersChangedEvent,
-        _: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.picker
-            .update(cx, |picker, _| picker.delegate.modifiers = ev.modifiers)
-    }
-
-    pub fn handle_new_worktree(
-        &mut self,
-        replace_current_window: bool,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.picker.update(cx, |picker, cx| {
-            let ix = picker.delegate.selected_index();
-            let Some(entry) = picker.delegate.matches.get(ix) else {
-                return;
-            };
-            let Some(default_branch) = picker.delegate.default_branch.clone() else {
-                return;
-            };
-            if !entry.is_new {
-                return;
-            }
-            picker.delegate.create_worktree(
-                entry.worktree.display_name(),
-                replace_current_window,
-                Some(default_branch.into()),
-                window,
-                cx,
-            );
-        })
-    }
-
-    pub fn handle_delete(
-        &mut self,
-        _: &DeleteWorktree,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.picker.update(cx, |picker, cx| {
-            picker
-                .delegate
-                .delete_at(picker.delegate.selected_index, window, cx)
-        })
-    }
 }
-impl ModalView for WorktreeList {}
-impl EventEmitter<DismissEvent> for WorktreeList {}
 
-impl Focusable for WorktreeList {
-    fn focus_handle(&self, _: &App) -> FocusHandle {
-        self.picker_focus_handle.clone()
+impl Focusable for WorktreePicker {
+    fn focus_handle(&self, _cx: &App) -> FocusHandle {
+        self.focus_handle.clone()
     }
 }
 
-impl Render for WorktreeList {
-    fn render(&mut self, _: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
+impl ModalView for WorktreePicker {}
+impl EventEmitter<DismissEvent> for WorktreePicker {}
+
+impl Render for WorktreePicker {
+    fn render(&mut self, _window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
         v_flex()
-            .key_context("GitWorktreeSelector")
-            .w(self.width)
-            .on_modifiers_changed(cx.listener(Self::handle_modifiers_changed))
-            .on_action(cx.listener(|this, _: &WorktreeFromDefault, w, cx| {
-                this.handle_new_worktree(false, w, cx)
-            }))
-            .on_action(cx.listener(|this, _: &WorktreeFromDefaultOnWindow, w, cx| {
-                this.handle_new_worktree(true, w, cx)
+            .key_context("WorktreePicker")
+            .w(rems(34.))
+            .elevation_3(cx)
+            .child(self.picker.clone())
+            .on_mouse_down_out(cx.listener(|_, _, _, cx| {
+                cx.emit(DismissEvent);
             }))
             .on_action(cx.listener(|this, _: &DeleteWorktree, window, cx| {
-                this.handle_delete(&DeleteWorktree, window, cx)
+                this.picker.update(cx, |picker, cx| {
+                    let ix = picker.delegate.selected_index;
+                    picker.delegate.delete_worktree(ix, window, cx);
+                });
             }))
-            .child(self.picker.clone())
-            .when(!self.embedded, |el| {
-                el.on_mouse_down_out({
-                    cx.listener(move |this, _, window, cx| {
-                        this.picker.update(cx, |this, cx| {
-                            this.cancel(&Default::default(), window, cx);
-                        })
-                    })
-                })
-            })
     }
 }
 
-#[derive(Debug, Clone)]
-struct WorktreeEntry {
-    worktree: GitWorktree,
-    positions: Vec<usize>,
-    is_new: bool,
+#[derive(Clone)]
+pub enum WorktreeEntry {
+    CreateFromCurrentBranch,
+    CreateFromDefaultBranch {
+        default_branch_name: String,
+    },
+    Separator,
+    Worktree {
+        worktree: GitWorktree,
+        positions: Vec<usize>,
+    },
+    CreateNamed {
+        name: String,
+        from_branch: Option<String>,
+        disabled_reason: Option<String>,
+    },
 }
 
-impl WorktreeEntry {
-    fn can_delete(&self, forbidden_deletion_path: Option<&PathBuf>) -> bool {
-        !self.is_new
-            && !self.worktree.is_main
-            && forbidden_deletion_path != Some(&self.worktree.path)
-    }
-}
-
-pub struct WorktreeListDelegate {
+pub struct WorktreePickerDelegate {
     matches: Vec<WorktreeEntry>,
-    all_worktrees: Option<Vec<GitWorktree>>,
-    workspace: WeakEntity<Workspace>,
-    repo: Option<Entity<Repository>>,
+    all_worktrees: Vec<GitWorktree>,
+    project_worktree_paths: HashSet<PathBuf>,
     selected_index: usize,
-    last_query: String,
-    modifiers: Modifiers,
+    project: Entity<Project>,
+    current_branch_name: Option<String>,
+    default_branch_name: Option<String>,
+    has_multiple_repositories: bool,
     focus_handle: FocusHandle,
-    default_branch: Option<SharedString>,
-    forbidden_deletion_path: Option<PathBuf>,
-    current_worktree_path: Option<PathBuf>,
+    show_footer: bool,
 }
 
-impl WorktreeListDelegate {
-    fn new(
-        workspace: WeakEntity<Workspace>,
-        repo: Option<Entity<Repository>>,
-        _window: &mut Window,
-        cx: &mut Context<WorktreeList>,
-    ) -> Self {
-        let current_worktree_path = repo
-            .as_ref()
-            .map(|r| r.read(cx).work_directory_abs_path.to_path_buf());
+impl WorktreePickerDelegate {
+    fn build_fixed_entries(&self) -> Vec<WorktreeEntry> {
+        let mut entries = Vec::new();
 
-        Self {
-            matches: vec![],
-            all_worktrees: None,
-            workspace,
-            selected_index: 0,
-            repo,
-            last_query: Default::default(),
-            modifiers: Default::default(),
-            focus_handle: cx.focus_handle(),
-            default_branch: None,
-            forbidden_deletion_path: None,
-            current_worktree_path,
-        }
-    }
+        entries.push(WorktreeEntry::CreateFromCurrentBranch);
 
-    fn create_worktree(
-        &self,
-        worktree_branch: &str,
-        replace_current_window: bool,
-        commit: Option<String>,
-        window: &mut Window,
-        cx: &mut Context<Picker<Self>>,
-    ) {
-        let Some(repo) = self.repo.clone() else {
-            return;
-        };
-
-        let branch = worktree_branch.to_string();
-        let workspace = self.workspace.clone();
-        cx.spawn_in(window, async move |_, cx| {
-            let (receiver, new_worktree_path) = repo.update(cx, |repo, cx| {
-                let worktree_directory_setting = ProjectSettings::get_global(cx)
-                    .git
-                    .worktree_directory
-                    .clone();
-                let new_worktree_path =
-                    repo.path_for_new_linked_worktree(&branch, &worktree_directory_setting)?;
-                let receiver = repo.create_worktree(
-                    git::repository::CreateWorktreeTarget::NewBranch {
-                        branch_name: branch.clone(),
-                        base_sha: commit,
-                    },
-                    new_worktree_path.clone(),
-                );
-                anyhow::Ok((receiver, new_worktree_path))
-            })?;
-            receiver.await??;
-
-            workspace.update(cx, |workspace, cx| {
-                if let Some(trusted_worktrees) = TrustedWorktrees::try_get_global(cx) {
-                    let repo_path = &repo.read(cx).snapshot().work_directory_abs_path;
-                    let project = workspace.project();
-                    if let Some((parent_worktree, _)) =
-                        project.read(cx).find_worktree(repo_path, cx)
-                    {
-                        let worktree_store = project.read(cx).worktree_store();
-                        trusted_worktrees.update(cx, |trusted_worktrees, cx| {
-                            if trusted_worktrees.can_trust(
-                                &worktree_store,
-                                parent_worktree.read(cx).id(),
-                                cx,
-                            ) {
-                                trusted_worktrees.trust(
-                                    &worktree_store,
-                                    HashSet::from_iter([PathTrust::AbsPath(
-                                        new_worktree_path.clone(),
-                                    )]),
-                                    cx,
-                                );
-                            }
-                        });
-                    }
+        if !self.has_multiple_repositories {
+            if let Some(ref default_branch) = self.default_branch_name {
+                let is_different = self
+                    .current_branch_name
+                    .as_ref()
+                    .is_none_or(|current| current != default_branch);
+                if is_different {
+                    entries.push(WorktreeEntry::CreateFromDefaultBranch {
+                        default_branch_name: default_branch.clone(),
+                    });
                 }
-            })?;
-
-            let (connection_options, app_state, is_local) =
-                workspace.update(cx, |workspace, cx| {
-                    let project = workspace.project().clone();
-                    let connection_options = project.read(cx).remote_connection_options(cx);
-                    let app_state = workspace.app_state().clone();
-                    let is_local = project.read(cx).is_local();
-                    (connection_options, app_state, is_local)
-                })?;
-
-            if is_local {
-                workspace
-                    .update_in(cx, |workspace, window, cx| {
-                        workspace.open_workspace_for_paths(
-                            OpenMode::Activate,
-                            vec![new_worktree_path],
-                            window,
-                            cx,
-                        )
-                    })?
-                    .await?;
-            } else if let Some(connection_options) = connection_options {
-                open_remote_worktree(
-                    connection_options,
-                    vec![new_worktree_path],
-                    app_state,
-                    workspace.clone(),
-                    replace_current_window,
-                    cx,
-                )
-                .await?;
             }
+        }
 
-            anyhow::Ok(())
-        })
-        .detach_and_prompt_err("Failed to create worktree", window, cx, |e, _, _| {
-            let msg = e.to_string();
-            if msg.contains("git.worktree_directory") {
-                Some(format!("Invalid git.worktree_directory setting: {}", e))
-            } else {
-                Some(msg)
-            }
-        });
+        entries
     }
 
-    fn open_worktree(
-        &self,
-        worktree_path: &PathBuf,
-        replace_current_window: bool,
-        window: &mut Window,
-        cx: &mut Context<Picker<Self>>,
-    ) {
-        let workspace = self.workspace.clone();
-        let path = worktree_path.clone();
-
-        let Some((connection_options, app_state, is_local)) = workspace
-            .update(cx, |workspace, cx| {
-                let project = workspace.project().clone();
-                let connection_options = project.read(cx).remote_connection_options(cx);
-                let app_state = workspace.app_state().clone();
-                let is_local = project.read(cx).is_local();
-                (connection_options, app_state, is_local)
-            })
-            .log_err()
-        else {
-            return;
-        };
-        let open_mode = if replace_current_window {
-            OpenMode::Activate
+    fn all_repo_worktrees(&self) -> &[GitWorktree] {
+        if self.has_multiple_repositories {
+            &[]
         } else {
-            OpenMode::NewWindow
-        };
-
-        if is_local {
-            let open_task = workspace.update(cx, |workspace, cx| {
-                workspace.open_workspace_for_paths(open_mode, vec![path], window, cx)
-            });
-            cx.spawn(async move |_, _| {
-                open_task?.await?;
-                anyhow::Ok(())
-            })
-            .detach_and_prompt_err(
-                "Failed to open worktree",
-                window,
-                cx,
-                |e, _, _| Some(e.to_string()),
-            );
-        } else if let Some(connection_options) = connection_options {
-            cx.spawn_in(window, async move |_, cx| {
-                open_remote_worktree(
-                    connection_options,
-                    vec![path],
-                    app_state,
-                    workspace,
-                    replace_current_window,
-                    cx,
-                )
-                .await
-            })
-            .detach_and_prompt_err(
-                "Failed to open worktree",
-                window,
-                cx,
-                |e, _, _| Some(e.to_string()),
-            );
+            &self.all_worktrees
         }
-
-        cx.emit(DismissEvent);
     }
 
-    fn base_branch<'a>(&'a self, cx: &'a mut Context<Picker<Self>>) -> Option<&'a str> {
-        self.repo
-            .as_ref()
-            .and_then(|repo| repo.read(cx).branch.as_ref().map(|b| b.name()))
+    fn can_delete_worktree(&self, worktree: &GitWorktree) -> bool {
+        !worktree.is_main && !self.project_worktree_paths.contains(&worktree.path)
     }
 
-    fn delete_at(&self, idx: usize, window: &mut Window, cx: &mut Context<Picker<Self>>) {
-        let Some(entry) = self.matches.get(idx).cloned() else {
+    fn delete_worktree(&self, ix: usize, window: &mut Window, cx: &mut Context<Picker<Self>>) {
+        let Some(entry) = self.matches.get(ix) else {
             return;
         };
-        if !entry.can_delete(self.forbidden_deletion_path.as_ref()) {
+        let WorktreeEntry::Worktree { worktree, .. } = entry else {
+            return;
+        };
+        if !self.can_delete_worktree(worktree) {
             return;
         }
-        let Some(repo) = self.repo.clone() else {
+
+        let repo = self.project.read(cx).active_repository(cx);
+        let Some(repo) = repo else {
             return;
         };
-        let workspace = self.workspace.clone();
-        let path = entry.worktree.path;
+        let path = worktree.path.clone();
 
         cx.spawn_in(window, async move |picker, cx| {
-            let result = repo
-                .update(cx, |repo, _| repo.remove_worktree(path.clone(), false))
-                .await?;
+            let receiver = repo.update(cx, |repo, _| repo.remove_worktree(path.clone(), false));
+            receiver
+                .await
+                .map_err(|_| anyhow::anyhow!("worktree removal was cancelled"))?
+                .map_err(|e| anyhow::anyhow!("failed to remove worktree: {e}"))?;
 
-            if let Err(e) = result {
-                log::error!("Failed to remove worktree: {}", e);
-                if let Some(workspace) = workspace.upgrade() {
-                    cx.update(|_window, cx| {
-                        show_error_toast(
-                            workspace,
-                            format!("worktree remove {}", path.display()),
-                            e,
-                            cx,
-                        )
-                    })?;
-                }
-                return Ok(());
-            }
-
-            picker.update_in(cx, |picker, _, cx| {
-                picker.delegate.matches.retain(|e| e.worktree.path != path);
-                if let Some(all_worktrees) = &mut picker.delegate.all_worktrees {
-                    all_worktrees.retain(|w| w.path != path);
-                }
-                picker.delegate.refresh_forbidden_deletion_path(cx);
+            picker.update_in(cx, |picker, _window, cx| {
+                picker.delegate.matches.retain(|e| {
+                    !matches!(e, WorktreeEntry::Worktree { worktree, .. } if worktree.path == path)
+                });
+                picker.delegate.all_worktrees.retain(|w| w.path != path);
                 if picker.delegate.matches.is_empty() {
                     picker.delegate.selected_index = 0;
                 } else if picker.delegate.selected_index >= picker.delegate.matches.len() {
@@ -555,139 +318,37 @@ impl WorktreeListDelegate {
 
             anyhow::Ok(())
         })
-        .detach();
+        .detach_and_log_err(cx);
     }
 
-    fn refresh_forbidden_deletion_path(&mut self, cx: &App) {
-        let Some(workspace) = self.workspace.upgrade() else {
-            debug_panic!("Workspace should always be available or else the picker would be closed");
-            self.forbidden_deletion_path = None;
+    fn sync_selected_index(&mut self, has_query: bool) {
+        if !has_query {
             return;
-        };
+        }
 
-        let visible_worktree_paths = workspace.read_with(cx, |workspace, cx| {
-            workspace
-                .project()
-                .read(cx)
-                .visible_worktrees(cx)
-                .map(|worktree| worktree.read(cx).abs_path().to_path_buf())
-                .collect::<Vec<_>>()
-        });
-
-        self.forbidden_deletion_path = if visible_worktree_paths.len() == 1 {
-            visible_worktree_paths.into_iter().next()
+        if let Some(index) = self
+            .matches
+            .iter()
+            .position(|entry| matches!(entry, WorktreeEntry::Worktree { .. }))
+        {
+            self.selected_index = index;
+        } else if let Some(index) = self
+            .matches
+            .iter()
+            .position(|entry| matches!(entry, WorktreeEntry::CreateNamed { .. }))
+        {
+            self.selected_index = index;
         } else {
-            None
-        };
+            self.selected_index = 0;
+        }
     }
 }
 
-async fn open_remote_worktree(
-    connection_options: RemoteConnectionOptions,
-    paths: Vec<PathBuf>,
-    app_state: Arc<workspace::AppState>,
-    workspace: WeakEntity<Workspace>,
-    replace_current_window: bool,
-    cx: &mut AsyncWindowContext,
-) -> anyhow::Result<()> {
-    let workspace_window = cx
-        .window_handle()
-        .downcast::<MultiWorkspace>()
-        .ok_or_else(|| anyhow::anyhow!("Window is not a Workspace window"))?;
-
-    let connect_task = workspace.update_in(cx, |workspace, window, cx| {
-        workspace.toggle_modal(window, cx, |window, cx| {
-            RemoteConnectionModal::new(&connection_options, Vec::new(), window, cx)
-        });
-
-        let prompt = workspace
-            .active_modal::<RemoteConnectionModal>(cx)
-            .expect("Modal just created")
-            .read(cx)
-            .prompt
-            .clone();
-
-        connect(
-            ConnectionIdentifier::setup(),
-            connection_options.clone(),
-            prompt,
-            window,
-            cx,
-        )
-        .prompt_err("Failed to connect", window, cx, |_, _, _| None)
-    })?;
-
-    let session = connect_task.await;
-
-    workspace
-        .update_in(cx, |workspace, _window, cx| {
-            if let Some(prompt) = workspace.active_modal::<RemoteConnectionModal>(cx) {
-                prompt.update(cx, |prompt, cx| prompt.finished(cx))
-            }
-        })
-        .ok();
-
-    let Some(Some(session)) = session else {
-        return Ok(());
-    };
-
-    let new_project: Entity<project::Project> = cx.update(|_, cx| {
-        project::Project::remote(
-            session,
-            app_state.client.clone(),
-            app_state.node_runtime.clone(),
-            app_state.user_store.clone(),
-            app_state.languages.clone(),
-            app_state.fs.clone(),
-            true,
-            cx,
-        )
-    })?;
-
-    let window_to_use = if replace_current_window {
-        workspace_window
-    } else {
-        let workspace_position = cx
-            .update(|_, cx| {
-                workspace::remote_workspace_position_from_db(connection_options.clone(), &paths, cx)
-            })?
-            .await
-            .context("fetching workspace position from db")?;
-
-        let mut options =
-            cx.update(|_, cx| (app_state.build_window_options)(workspace_position.display, cx))?;
-        options.window_bounds = workspace_position.window_bounds;
-
-        cx.open_window(options, |window, cx| {
-            let workspace = cx.new(|cx| {
-                let mut workspace =
-                    Workspace::new(None, new_project.clone(), app_state.clone(), window, cx);
-                workspace.centered_layout = workspace_position.centered_layout;
-                workspace
-            });
-            cx.new(|cx| MultiWorkspace::new(workspace, window, cx))
-        })?
-    };
-
-    workspace::open_remote_project_with_existing_connection(
-        connection_options,
-        new_project,
-        paths,
-        app_state,
-        window_to_use,
-        None,
-        cx,
-    )
-    .await?;
-
-    Ok(())
-}
-
-impl PickerDelegate for WorktreeListDelegate {
-    type ListItem = ListItem;
+impl PickerDelegate for WorktreePickerDelegate {
+    type ListItem = AnyElement;
 
     fn placeholder_text(&self, _window: &mut Window, _cx: &mut App) -> Arc<str> {
-        "Select worktree…".into()
+        "Select a worktree…".into()
     }
 
     fn editor_position(&self) -> PickerEditorPosition {
@@ -706,9 +367,13 @@ impl PickerDelegate for WorktreeListDelegate {
         &mut self,
         ix: usize,
         _window: &mut Window,
-        _: &mut Context<Picker<Self>>,
+        _cx: &mut Context<Picker<Self>>,
     ) {
         self.selected_index = ix;
+    }
+
+    fn can_select(&self, ix: usize, _window: &mut Window, _cx: &mut Context<Picker<Self>>) -> bool {
+        !matches!(self.matches.get(ix), Some(WorktreeEntry::Separator))
     }
 
     fn update_matches(
@@ -717,104 +382,232 @@ impl PickerDelegate for WorktreeListDelegate {
         window: &mut Window,
         cx: &mut Context<Picker<Self>>,
     ) -> Task<()> {
-        let Some(all_worktrees) = self.all_worktrees.clone() else {
-            return Task::ready(());
+        let repo_worktrees = self.all_repo_worktrees().to_vec();
+
+        let normalized_query = query.replace(' ', "-");
+        let main_worktree_path = self
+            .all_worktrees
+            .iter()
+            .find(|wt| wt.is_main)
+            .map(|wt| wt.path.clone());
+        let has_named_worktree = self.all_worktrees.iter().any(|worktree| {
+            worktree.directory_name(main_worktree_path.as_deref()) == normalized_query
+        });
+        let create_named_disabled_reason: Option<String> = if self.has_multiple_repositories {
+            Some("Cannot create a named worktree in a project with multiple repositories".into())
+        } else if has_named_worktree {
+            Some("A worktree with this name already exists".into())
+        } else {
+            None
         };
 
-        cx.spawn_in(window, async move |picker, cx| {
-            let main_worktree_path = all_worktrees
-                .iter()
-                .find(|wt| wt.is_main)
-                .map(|wt| wt.path.clone());
+        let show_default_branch_create = !self.has_multiple_repositories
+            && self.default_branch_name.as_ref().is_some_and(|default| {
+                self.current_branch_name
+                    .as_ref()
+                    .is_none_or(|current| current != default)
+            });
+        let default_branch_name = self.default_branch_name.clone();
 
-            let mut matches: Vec<WorktreeEntry> = if query.is_empty() {
-                all_worktrees
-                    .into_iter()
-                    .map(|worktree| WorktreeEntry {
+        if query.is_empty() {
+            let mut matches = self.build_fixed_entries();
+
+            if !repo_worktrees.is_empty() {
+                let main_worktree_path = repo_worktrees
+                    .iter()
+                    .find(|wt| wt.is_main)
+                    .map(|wt| wt.path.clone());
+
+                let mut sorted = repo_worktrees;
+                let project_paths = &self.project_worktree_paths;
+
+                sorted.sort_by(|a, b| {
+                    let a_is_current = project_paths.contains(&a.path);
+                    let b_is_current = project_paths.contains(&b.path);
+                    b_is_current.cmp(&a_is_current).then_with(|| {
+                        a.directory_name(main_worktree_path.as_deref())
+                            .cmp(&b.directory_name(main_worktree_path.as_deref()))
+                    })
+                });
+
+                matches.push(WorktreeEntry::Separator);
+                for worktree in sorted {
+                    matches.push(WorktreeEntry::Worktree {
                         worktree,
                         positions: Vec::new(),
-                        is_new: false,
-                    })
-                    .collect()
-            } else {
-                let candidates = all_worktrees
-                    .iter()
-                    .enumerate()
-                    .map(|(ix, worktree)| {
-                        StringMatchCandidate::new(
-                            ix,
-                            &worktree.directory_name(main_worktree_path.as_deref()),
-                        )
-                    })
-                    .collect::<Vec<StringMatchCandidate>>();
-                fuzzy::match_strings(
-                    &candidates,
-                    &query,
-                    true,
-                    true,
-                    10000,
-                    &Default::default(),
-                    cx.background_executor().clone(),
+                    });
+                }
+            }
+
+            self.matches = matches;
+            self.sync_selected_index(false);
+            return Task::ready(());
+        }
+
+        let main_worktree_path = repo_worktrees
+            .iter()
+            .find(|wt| wt.is_main)
+            .map(|wt| wt.path.clone());
+        let candidates: Vec<_> = repo_worktrees
+            .iter()
+            .enumerate()
+            .map(|(ix, worktree)| {
+                StringMatchCandidate::new(
+                    ix,
+                    &worktree.directory_name(main_worktree_path.as_deref()),
                 )
-                .await
-                .into_iter()
-                .map(|candidate| WorktreeEntry {
-                    worktree: all_worktrees[candidate.candidate_id].clone(),
-                    positions: candidate.positions,
-                    is_new: false,
-                })
-                .collect()
-            };
+            })
+            .collect();
+
+        let executor = cx.background_executor().clone();
+
+        let task = cx.background_executor().spawn(async move {
+            fuzzy::match_strings(
+                &candidates,
+                &query,
+                true,
+                true,
+                10000,
+                &Default::default(),
+                executor,
+            )
+            .await
+        });
+
+        let repo_worktrees_clone = repo_worktrees;
+        cx.spawn_in(window, async move |picker, cx| {
+            let fuzzy_matches = task.await;
+
             picker
-                .update(cx, |picker, _| {
-                    if !query.is_empty()
-                        && !matches.first().is_some_and(|entry| {
-                            entry.worktree.directory_name(main_worktree_path.as_deref()) == query
-                        })
-                    {
-                        let query = query.replace(' ', "-");
-                        matches.push(WorktreeEntry {
-                            worktree: GitWorktree {
-                                path: Default::default(),
-                                ref_name: Some(format!("refs/heads/{query}").into()),
-                                sha: Default::default(),
-                                is_main: false,
-                                is_bare: false,
-                            },
-                            positions: Vec::new(),
-                            is_new: true,
-                        })
+                .update_in(cx, |picker, _window, cx| {
+                    let mut new_matches: Vec<WorktreeEntry> = Vec::new();
+
+                    for candidate in &fuzzy_matches {
+                        new_matches.push(WorktreeEntry::Worktree {
+                            worktree: repo_worktrees_clone[candidate.candidate_id].clone(),
+                            positions: candidate.positions.clone(),
+                        });
                     }
-                    let delegate = &mut picker.delegate;
-                    delegate.matches = matches;
-                    if delegate.matches.is_empty() {
-                        delegate.selected_index = 0;
-                    } else {
-                        delegate.selected_index =
-                            core::cmp::min(delegate.selected_index, delegate.matches.len() - 1);
+
+                    if !new_matches.is_empty() {
+                        new_matches.push(WorktreeEntry::Separator);
                     }
-                    delegate.last_query = query;
+                    new_matches.push(WorktreeEntry::CreateNamed {
+                        name: normalized_query.clone(),
+                        from_branch: None,
+                        disabled_reason: create_named_disabled_reason.clone(),
+                    });
+                    if show_default_branch_create {
+                        if let Some(ref default_branch) = default_branch_name {
+                            new_matches.push(WorktreeEntry::CreateNamed {
+                                name: normalized_query.clone(),
+                                from_branch: Some(default_branch.clone()),
+                                disabled_reason: create_named_disabled_reason.clone(),
+                            });
+                        }
+                    }
+
+                    picker.delegate.matches = new_matches;
+                    picker.delegate.sync_selected_index(true);
+
+                    cx.notify();
                 })
                 .log_err();
         })
     }
 
     fn confirm(&mut self, secondary: bool, window: &mut Window, cx: &mut Context<Picker<Self>>) {
-        let Some(entry) = self.matches.get(self.selected_index()) else {
+        let Some(entry) = self.matches.get(self.selected_index) else {
             return;
         };
-        if entry.is_new {
-            self.create_worktree(&entry.worktree.display_name(), secondary, None, window, cx);
-        } else {
-            self.open_worktree(&entry.worktree.path, !secondary, window, cx);
+
+        match entry {
+            WorktreeEntry::Separator => return,
+
+            WorktreeEntry::CreateFromCurrentBranch => {
+                window.dispatch_action(
+                    Box::new(CreateWorktree {
+                        worktree_name: None,
+                        branch_target: NewWorktreeBranchTarget::CurrentBranch,
+                    }),
+                    cx,
+                );
+            }
+
+            WorktreeEntry::CreateFromDefaultBranch {
+                default_branch_name,
+            } => {
+                window.dispatch_action(
+                    Box::new(CreateWorktree {
+                        worktree_name: None,
+                        branch_target: NewWorktreeBranchTarget::ExistingBranch {
+                            name: default_branch_name.clone(),
+                        },
+                    }),
+                    cx,
+                );
+            }
+
+            WorktreeEntry::Worktree { worktree, .. } => {
+                let is_current = self.project_worktree_paths.contains(&worktree.path);
+
+                if !is_current {
+                    if secondary {
+                        window.dispatch_action(
+                            Box::new(OpenWorktreeInNewWindow {
+                                path: worktree.path.clone(),
+                            }),
+                            cx,
+                        );
+                    } else {
+                        let main_worktree_path = self
+                            .all_worktrees
+                            .iter()
+                            .find(|wt| wt.is_main)
+                            .map(|wt| wt.path.as_path());
+                        window.dispatch_action(
+                            Box::new(SwitchWorktree {
+                                path: worktree.path.clone(),
+                                display_name: worktree.directory_name(main_worktree_path),
+                            }),
+                            cx,
+                        );
+                    }
+                }
+            }
+
+            WorktreeEntry::CreateNamed {
+                name,
+                from_branch,
+                disabled_reason: None,
+            } => {
+                let branch_target = match from_branch {
+                    Some(branch) => NewWorktreeBranchTarget::ExistingBranch {
+                        name: branch.clone(),
+                    },
+                    None => NewWorktreeBranchTarget::CurrentBranch,
+                };
+                window.dispatch_action(
+                    Box::new(CreateWorktree {
+                        worktree_name: Some(name.clone()),
+                        branch_target,
+                    }),
+                    cx,
+                );
+            }
+
+            WorktreeEntry::CreateNamed {
+                disabled_reason: Some(_),
+                ..
+            } => {
+                return;
+            }
         }
 
         cx.emit(DismissEvent);
     }
 
-    fn dismissed(&mut self, _: &mut Window, cx: &mut Context<Picker<Self>>) {
-        cx.emit(DismissEvent);
-    }
+    fn dismissed(&mut self, _window: &mut Window, _cx: &mut Context<Picker<Self>>) {}
 
     fn render_match(
         &self,
@@ -823,83 +616,18 @@ impl PickerDelegate for WorktreeListDelegate {
         _window: &mut Window,
         cx: &mut Context<Picker<Self>>,
     ) -> Option<Self::ListItem> {
-        let entry = &self.matches.get(ix)?;
-        let path = entry.worktree.path.compact().to_string_lossy().to_string();
-        let sha = entry
-            .worktree
-            .sha
-            .clone()
-            .chars()
-            .take(7)
-            .collect::<String>();
+        let entry = self.matches.get(ix)?;
+        let project = self.project.read(cx);
+        let is_create_disabled = project.repositories(cx).is_empty() || project.is_via_collab();
 
-        let (branch_name, sublabel) = if entry.is_new {
-            (
-                Label::new(format!(
-                    "Create Worktree: \"{}\"…",
-                    entry.worktree.display_name()
-                ))
-                .truncate()
-                .into_any_element(),
-                format!(
-                    "based off {}",
-                    self.base_branch(cx).unwrap_or("the current branch")
-                ),
-            )
-        } else {
-            let main_worktree_path = self
-                .all_worktrees
-                .as_ref()
-                .and_then(|wts| wts.iter().find(|wt| wt.is_main))
-                .map(|wt| wt.path.as_path());
-            let display_name = entry.worktree.directory_name(main_worktree_path);
-            let first_line = display_name.lines().next().unwrap_or(&display_name);
-            let positions: Vec<_> = entry
-                .positions
-                .iter()
-                .copied()
-                .filter(|&pos| pos < first_line.len())
-                .collect();
+        let no_git_reason: SharedString = "Requires a Git repository in the project".into();
 
-            (
-                HighlightedLabel::new(first_line.to_owned(), positions)
-                    .truncate()
-                    .into_any_element(),
-                path,
-            )
-        };
-
-        let focus_handle = self.focus_handle.clone();
-
-        let can_delete = entry.can_delete(self.forbidden_deletion_path.as_ref());
-
-        let delete_button = |entry_ix: usize| {
-            IconButton::new(("delete-worktree", entry_ix), IconName::Trash)
-                .icon_size(IconSize::Small)
-                .tooltip(move |_, cx| {
-                    Tooltip::for_action_in("Delete Worktree", &DeleteWorktree, &focus_handle, cx)
-                })
-                .on_click(cx.listener(move |this, _, window, cx| {
-                    this.delegate.delete_at(entry_ix, window, cx);
-                }))
-        };
-
-        let is_current = !entry.is_new
-            && self
-                .current_worktree_path
-                .as_ref()
-                .is_some_and(|current| *current == entry.worktree.path);
-
-        let entry_icon = if entry.is_new {
-            IconName::Plus
-        } else if is_current {
-            IconName::Check
-        } else {
-            IconName::GitWorktree
-        };
-
-        Some(
-            ListItem::new(format!("worktree-menu-{ix}"))
+        let create_new_list_item = |id: SharedString,
+                                    label: SharedString,
+                                    disabled_tooltip: Option<SharedString>,
+                                    selected: bool| {
+            let is_disabled = disabled_tooltip.is_some();
+            ListItem::new(id)
                 .inset(true)
                 .spacing(ListItemSpacing::Sparse)
                 .toggle_state(selected)
@@ -908,114 +636,275 @@ impl PickerDelegate for WorktreeListDelegate {
                         .w_full()
                         .gap_2p5()
                         .child(
-                            Icon::new(entry_icon)
-                                .color(if is_current {
-                                    Color::Accent
-                                } else {
-                                    Color::Muted
+                            Icon::new(IconName::Plus)
+                                .map(|this| {
+                                    if is_disabled {
+                                        this.color(Color::Disabled)
+                                    } else {
+                                        this.color(Color::Muted)
+                                    }
                                 })
                                 .size(IconSize::Small),
                         )
-                        .child(v_flex().w_full().min_w_0().child(branch_name).map(|this| {
-                            if entry.is_new {
-                                this.child(
-                                    Label::new(sublabel)
-                                        .size(LabelSize::Small)
-                                        .color(Color::Muted)
-                                        .truncate(),
+                        .child(
+                            Label::new(label).when(is_disabled, |this| this.color(Color::Disabled)),
+                        ),
+                )
+                .when_some(disabled_tooltip, |this, reason| {
+                    this.tooltip(Tooltip::text(reason))
+                })
+                .into_any_element()
+        };
+
+        match entry {
+            WorktreeEntry::Separator => Some(
+                div()
+                    .py(DynamicSpacing::Base04.rems(cx))
+                    .child(Divider::horizontal())
+                    .into_any_element(),
+            ),
+
+            WorktreeEntry::CreateFromCurrentBranch => {
+                let branch_label = if self.has_multiple_repositories {
+                    "current branches".to_string()
+                } else {
+                    self.current_branch_name
+                        .clone()
+                        .unwrap_or_else(|| "HEAD".to_string())
+                };
+
+                let label = format!("Create new worktree based on {branch_label}");
+
+                let disabled_tooltip = is_create_disabled.then(|| no_git_reason.clone());
+
+                let item = create_new_list_item(
+                    "create-from-current".to_string().into(),
+                    label.into(),
+                    disabled_tooltip,
+                    selected,
+                );
+
+                Some(item.into_any_element())
+            }
+
+            WorktreeEntry::CreateFromDefaultBranch {
+                default_branch_name,
+            } => {
+                let label = format!("Create new worktree based on {default_branch_name}");
+
+                let disabled_tooltip = is_create_disabled.then(|| no_git_reason.clone());
+
+                let item = create_new_list_item(
+                    "create-from-main".to_string().into(),
+                    label.into(),
+                    disabled_tooltip,
+                    selected,
+                );
+
+                Some(item.into_any_element())
+            }
+
+            WorktreeEntry::Worktree {
+                worktree,
+                positions,
+            } => {
+                let main_worktree_path = self
+                    .all_worktrees
+                    .iter()
+                    .find(|wt| wt.is_main)
+                    .map(|wt| wt.path.as_path());
+                let display_name = worktree.directory_name(main_worktree_path);
+                let first_line = display_name.lines().next().unwrap_or(&display_name);
+                let positions: Vec<_> = positions
+                    .iter()
+                    .copied()
+                    .filter(|&pos| pos < first_line.len())
+                    .collect();
+                let path = worktree.path.compact().to_string_lossy().to_string();
+                let sha = worktree.sha.chars().take(7).collect::<String>();
+
+                let is_current = self.project_worktree_paths.contains(&worktree.path);
+                let can_delete = self.can_delete_worktree(worktree);
+
+                let entry_icon = if is_current {
+                    IconName::Check
+                } else {
+                    IconName::GitWorktree
+                };
+
+                Some(
+                    ListItem::new(SharedString::from(format!("worktree-{ix}")))
+                        .inset(true)
+                        .spacing(ListItemSpacing::Sparse)
+                        .toggle_state(selected)
+                        .child(
+                            h_flex()
+                                .w_full()
+                                .gap_2p5()
+                                .child(
+                                    Icon::new(entry_icon)
+                                        .color(if is_current {
+                                            Color::Accent
+                                        } else {
+                                            Color::Muted
+                                        })
+                                        .size(IconSize::Small),
                                 )
-                            } else {
-                                this.child(
-                                    h_flex()
+                                .child(
+                                    v_flex()
                                         .w_full()
                                         .min_w_0()
-                                        .gap_1p5()
-                                        .when_some(
-                                            entry.worktree.branch_name().map(|b| b.to_string()),
-                                            |this, branch| {
-                                                this.child(
-                                                    Label::new(branch)
-                                                        .size(LabelSize::Small)
-                                                        .color(Color::Muted),
+                                        .child(
+                                            HighlightedLabel::new(first_line.to_owned(), positions)
+                                                .truncate(),
+                                        )
+                                        .child(
+                                            h_flex()
+                                                .w_full()
+                                                .min_w_0()
+                                                .gap_1p5()
+                                                .when_some(
+                                                    worktree.branch_name().map(|b| b.to_string()),
+                                                    |this, branch| {
+                                                        this.child(
+                                                            Label::new(branch)
+                                                                .size(LabelSize::Small)
+                                                                .color(Color::Muted),
+                                                        )
+                                                        .child(
+                                                            Label::new("\u{2022}")
+                                                                .alpha(0.5)
+                                                                .color(Color::Muted)
+                                                                .size(LabelSize::Small),
+                                                        )
+                                                    },
                                                 )
+                                                .when(!sha.is_empty(), |this| {
+                                                    this.child(
+                                                        Label::new(sha)
+                                                            .size(LabelSize::Small)
+                                                            .color(Color::Muted),
+                                                    )
+                                                    .child(
+                                                        Label::new("\u{2022}")
+                                                            .alpha(0.5)
+                                                            .color(Color::Muted)
+                                                            .size(LabelSize::Small),
+                                                    )
+                                                })
                                                 .child(
-                                                    Label::new("•")
-                                                        .alpha(0.5)
+                                                    Label::new(path)
+                                                        .truncate_start()
                                                         .color(Color::Muted)
-                                                        .size(LabelSize::Small),
-                                                )
-                                            },
+                                                        .size(LabelSize::Small)
+                                                        .flex_1(),
+                                                ),
+                                        ),
+                                ),
+                        )
+                        .when(!is_current, |this| {
+                            let open_in_new_window_button =
+                                IconButton::new(("open-new-window", ix), IconName::ArrowUpRight)
+                                    .icon_size(IconSize::Small)
+                                    .tooltip(Tooltip::text("Open in New Window"))
+                                    .on_click(cx.listener(move |picker, _, window, cx| {
+                                        let Some(entry) = picker.delegate.matches.get(ix) else {
+                                            return;
+                                        };
+                                        if let WorktreeEntry::Worktree { worktree, .. } = entry {
+                                            window.dispatch_action(
+                                                Box::new(OpenWorktreeInNewWindow {
+                                                    path: worktree.path.clone(),
+                                                }),
+                                                cx,
+                                            );
+                                            cx.emit(DismissEvent);
+                                        }
+                                    }));
+
+                            let focus_handle_delete = self.focus_handle.clone();
+                            let delete_button =
+                                IconButton::new(("delete-worktree", ix), IconName::Trash)
+                                    .icon_size(IconSize::Small)
+                                    .tooltip(move |_, cx| {
+                                        Tooltip::for_action_in(
+                                            "Delete Worktree",
+                                            &DeleteWorktree,
+                                            &focus_handle_delete,
+                                            cx,
                                         )
-                                        .child(
-                                            Label::new(sha)
-                                                .size(LabelSize::Small)
-                                                .color(Color::Muted),
-                                        )
-                                        .child(
-                                            Label::new("•")
-                                                .alpha(0.5)
-                                                .color(Color::Muted)
-                                                .size(LabelSize::Small),
-                                        )
-                                        .child(
-                                            Label::new(sublabel)
-                                                .truncate_start()
-                                                .color(Color::Muted)
-                                                .size(LabelSize::Small)
-                                                .flex_1(),
-                                        )
-                                        .into_any_element(),
-                                )
-                            }
-                        })),
+                                    })
+                                    .on_click(cx.listener(move |picker, _, window, cx| {
+                                        picker.delegate.delete_worktree(ix, window, cx);
+                                    }));
+
+                            this.end_slot(
+                                h_flex()
+                                    .gap_0p5()
+                                    .child(open_in_new_window_button)
+                                    .when(can_delete, |this| this.child(delete_button)),
+                            )
+                            .show_end_slot_on_hover()
+                        })
+                        .into_any_element(),
                 )
-                .when(!entry.is_new && !is_current, |this| {
-                    let focus_handle = self.focus_handle.clone();
-                    let open_in_new_window_button =
-                        IconButton::new(("open-new-window", ix), IconName::ArrowUpRight)
-                            .icon_size(IconSize::Small)
-                            .tooltip(move |_, cx| {
-                                Tooltip::for_action_in(
-                                    "Open in New Window",
-                                    &menu::SecondaryConfirm,
-                                    &focus_handle,
-                                    cx,
-                                )
-                            })
-                            .on_click(|_, window, cx| {
-                                window.dispatch_action(menu::SecondaryConfirm.boxed_clone(), cx);
-                            });
+            }
 
-                    this.end_slot(
-                        h_flex()
-                            .gap_0p5()
-                            .child(open_in_new_window_button)
-                            .when(can_delete, |this| this.child(delete_button(ix))),
-                    )
-                    .show_end_slot_on_hover()
-                }),
-        )
-    }
+            WorktreeEntry::CreateNamed {
+                name,
+                from_branch,
+                disabled_reason,
+            } => {
+                let branch_label = from_branch
+                    .as_deref()
+                    .unwrap_or(self.current_branch_name.as_deref().unwrap_or("HEAD"));
+                let label = format!("Create \"{name}\" based on {branch_label}");
+                let element_id = match from_branch {
+                    Some(branch) => format!("create-named-from-{branch}"),
+                    None => "create-named-from-current".to_string(),
+                };
 
-    fn no_matches_text(&self, _window: &mut Window, _cx: &mut App) -> Option<SharedString> {
-        Some("No worktrees found".into())
+                let item = create_new_list_item(
+                    element_id.into(),
+                    label.into(),
+                    disabled_reason.clone().map(SharedString::from),
+                    selected,
+                );
+
+                Some(item.into_any_element())
+            }
+        }
     }
 
     fn render_footer(&self, _: &mut Window, cx: &mut Context<Picker<Self>>) -> Option<AnyElement> {
+        if !self.show_footer {
+            return None;
+        }
+
         let focus_handle = self.focus_handle.clone();
         let selected_entry = self.matches.get(self.selected_index);
-        let is_creating = selected_entry.is_some_and(|entry| entry.is_new);
-        let can_delete = selected_entry
-            .is_some_and(|entry| entry.can_delete(self.forbidden_deletion_path.as_ref()));
-        let is_current = selected_entry.is_some_and(|entry| {
-            !entry.is_new
-                && self
-                    .current_worktree_path
-                    .as_ref()
-                    .is_some_and(|current| *current == entry.worktree.path)
+
+        let is_creating = selected_entry.is_some_and(|e| {
+            matches!(
+                e,
+                WorktreeEntry::CreateFromCurrentBranch
+                    | WorktreeEntry::CreateFromDefaultBranch { .. }
+                    | WorktreeEntry::CreateNamed { .. }
+            )
         });
 
-        let footer_container = h_flex()
+        let is_existing_worktree =
+            selected_entry.is_some_and(|e| matches!(e, WorktreeEntry::Worktree { .. }));
+
+        let can_delete = selected_entry.is_some_and(|e| {
+            matches!(e, WorktreeEntry::Worktree { worktree, .. } if self.can_delete_worktree(worktree))
+        });
+
+        let is_current = selected_entry.is_some_and(|e| {
+            matches!(e, WorktreeEntry::Worktree { worktree, .. } if self.project_worktree_paths.contains(&worktree.path))
+        });
+
+        let footer = h_flex()
             .w_full()
             .p_1p5()
             .gap_0p5()
@@ -1024,44 +913,25 @@ impl PickerDelegate for WorktreeListDelegate {
             .border_color(cx.theme().colors().border_variant);
 
         if is_creating {
-            let from_default_button = self.default_branch.as_ref().map(|default_branch| {
-                Button::new(
-                    "worktree-from-default",
-                    format!("Create from: {default_branch}"),
-                )
-                .key_binding(
-                    KeyBinding::for_action_in(&WorktreeFromDefault, &focus_handle, cx)
-                        .map(|kb| kb.size(rems_from_px(12.))),
-                )
-                .on_click(|_, window, cx| {
-                    window.dispatch_action(WorktreeFromDefault.boxed_clone(), cx)
-                })
-            });
-
-            let current_branch = self.base_branch(cx).unwrap_or("current branch");
-
             Some(
-                footer_container
-                    .when_some(from_default_button, |this, button| this.child(button))
+                footer
                     .child(
-                        Button::new(
-                            "worktree-from-current",
-                            format!("Create from: {current_branch}"),
-                        )
-                        .key_binding(
-                            KeyBinding::for_action_in(&menu::Confirm, &focus_handle, cx)
-                                .map(|kb| kb.size(rems_from_px(12.))),
-                        )
-                        .on_click(|_, window, cx| {
-                            window.dispatch_action(menu::Confirm.boxed_clone(), cx)
-                        }),
+                        Button::new("create-worktree", "Create")
+                            .key_binding(
+                                KeyBinding::for_action_in(&menu::Confirm, &focus_handle, cx)
+                                    .map(|kb| kb.size(rems_from_px(12.))),
+                            )
+                            .on_click(|_, window, cx| {
+                                window.dispatch_action(menu::Confirm.boxed_clone(), cx)
+                            }),
                     )
                     .into_any(),
             )
-        } else {
+        } else if is_existing_worktree {
             Some(
-                footer_container
+                footer
                     .when(can_delete, |this| {
+                        let focus_handle = focus_handle.clone();
                         this.child(
                             Button::new("delete-worktree", "Delete")
                                 .key_binding(
@@ -1075,22 +945,15 @@ impl PickerDelegate for WorktreeListDelegate {
                     })
                     .when(!is_current, |this| {
                         this.child(
-                            Button::new("open-in-new-window", "Open in New Window")
-                                .key_binding(
-                                    KeyBinding::for_action_in(
-                                        &menu::SecondaryConfirm,
-                                        &focus_handle,
-                                        cx,
-                                    )
-                                    .map(|kb| kb.size(rems_from_px(12.))),
-                                )
-                                .on_click(|_, window, cx| {
+                            Button::new("open-in-new-window", "Open in New Window").on_click(
+                                |_, window, cx| {
                                     window.dispatch_action(menu::SecondaryConfirm.boxed_clone(), cx)
-                                }),
+                                },
+                            ),
                         )
                     })
                     .child(
-                        Button::new("open-in-window", "Open")
+                        Button::new("open-worktree", "Open")
                             .key_binding(
                                 KeyBinding::for_action_in(&menu::Confirm, &focus_handle, cx)
                                     .map(|kb| kb.size(rems_from_px(12.))),
@@ -1101,6 +964,106 @@ impl PickerDelegate for WorktreeListDelegate {
                     )
                     .into_any(),
             )
+        } else {
+            None
         }
     }
+}
+
+pub async fn open_remote_worktree(
+    connection_options: remote::RemoteConnectionOptions,
+    paths: Vec<PathBuf>,
+    app_state: Arc<workspace::AppState>,
+    workspace: gpui::WeakEntity<Workspace>,
+    cx: &mut gpui::AsyncWindowContext,
+) -> anyhow::Result<()> {
+    let connect_task = workspace.update_in(cx, |workspace, window, cx| {
+        workspace.toggle_modal(window, cx, |window, cx| {
+            remote_connection::RemoteConnectionModal::new(
+                &connection_options,
+                Vec::new(),
+                window,
+                cx,
+            )
+        });
+
+        let prompt = workspace
+            .active_modal::<remote_connection::RemoteConnectionModal>(cx)
+            .expect("Modal just created")
+            .read(cx)
+            .prompt
+            .clone();
+
+        remote_connection::connect(
+            remote::remote_client::ConnectionIdentifier::setup(),
+            connection_options.clone(),
+            prompt,
+            window,
+            cx,
+        )
+        .prompt_err("Failed to connect", window, cx, |_, _, _| None)
+    })?;
+
+    let session = connect_task.await;
+
+    workspace
+        .update_in(cx, |workspace, _window, cx| {
+            if let Some(prompt) =
+                workspace.active_modal::<remote_connection::RemoteConnectionModal>(cx)
+            {
+                prompt.update(cx, |prompt, cx| prompt.finished(cx))
+            }
+        })
+        .ok();
+
+    let Some(Some(session)) = session else {
+        return Ok(());
+    };
+
+    let new_project = cx.update(|_, cx| {
+        project::Project::remote(
+            session,
+            app_state.client.clone(),
+            app_state.node_runtime.clone(),
+            app_state.user_store.clone(),
+            app_state.languages.clone(),
+            app_state.fs.clone(),
+            true,
+            cx,
+        )
+    })?;
+
+    let workspace_position = cx
+        .update(|_, cx| {
+            workspace::remote_workspace_position_from_db(connection_options.clone(), &paths, cx)
+        })?
+        .await
+        .context("fetching workspace position from db")?;
+
+    let mut options =
+        cx.update(|_, cx| (app_state.build_window_options)(workspace_position.display, cx))?;
+    options.window_bounds = workspace_position.window_bounds;
+
+    let new_window = cx.open_window(options, |window, cx| {
+        let workspace = cx.new(|cx| {
+            let mut workspace =
+                Workspace::new(None, new_project.clone(), app_state.clone(), window, cx);
+            workspace.centered_layout = workspace_position.centered_layout;
+            workspace
+        });
+        cx.new(|cx| MultiWorkspace::new(workspace, window, cx))
+    })?;
+
+    workspace::open_remote_project_with_existing_connection(
+        connection_options,
+        new_project,
+        paths,
+        app_state,
+        new_window,
+        None,
+        cx,
+    )
+    .await?;
+
+    Ok(())
 }
