+++ crates/title_bar/src/title_bar.rs
@@ -7,6 +7,7 @@ mod update_version;
 
 use crate::application_menu::{ApplicationMenu, show_menus};
 use crate::plan_chip::PlanChip;
+use git_ui::worktree_picker::WorktreePicker;
 pub use platform_title_bar::{
     self, DraggedWindowTab, MergeAllWindows, MoveTabToNewWindow, PlatformTitleBar,
     ShowNextWindowTab, ShowPreviousWindowTab,
@@ -816,11 +817,13 @@ impl TitleBar {
         repository: Entity<project::git_store::Repository>,
         linked_worktree_name: Option<SharedString>,
         cx: &mut Context<Self>,
-    ) -> Option<impl IntoElement> {
+    ) -> Option<AnyElement> {
         let workspace = self.workspace.upgrade()?;
 
-        let (branch_name, icon_info) = {
+        let (branch_name, icon_info, is_detached_head) = {
             let repo = repository.read(cx);
+
+            let is_detached_head = repo.branch.is_none();
 
             let branch_name = repo
                 .branch
@@ -851,67 +854,109 @@ impl TitleBar {
                 (IconName::GitBranch, Color::Muted)
             };
 
-            (branch_name, icon_info)
+            (branch_name, icon_info, is_detached_head)
         };
 
         let branch_name = branch_name?;
         let settings = TitleBarSettings::get_global(cx);
         let effective_repository = Some(repository);
 
-        Some(
-            PopoverMenu::new("branch-menu")
+        let worktree_label: SharedString = linked_worktree_name.unwrap_or_else(|| "main".into());
+
+        let worktree_tooltip_label = worktree_label.clone();
+
+        let worktree_button = {
+            let project = self.project.clone();
+            PopoverMenu::new("worktree-picker-menu")
                 .menu(move |window, cx| {
-                    Some(git_ui::git_picker::popover(
-                        workspace.downgrade(),
-                        effective_repository.clone(),
-                        git_ui::git_picker::GitPickerTab::Branches,
-                        gpui::rems(34.),
-                        window,
-                        cx,
-                    ))
+                    Some(cx.new(|cx| WorktreePicker::new(project.clone(), window, cx)))
                 })
                 .trigger_with_tooltip(
-                    ButtonLike::new("project_branch_trigger")
+                    ButtonLike::new("worktree_picker_trigger")
                         .selected_style(ButtonStyle::Tinted(TintColor::Accent))
                         .child(
                             h_flex()
                                 .gap_0p5()
-                                .when(settings.show_branch_icon, |this| {
-                                    let (icon, icon_color) = icon_info;
-                                    this.child(
-                                        Icon::new(icon).size(IconSize::XSmall).color(icon_color),
-                                    )
-                                })
-                                .when_some(linked_worktree_name.as_ref(), |this, worktree_name| {
-                                    this.child(
-                                        Label::new(worktree_name)
-                                            .size(LabelSize::Small)
-                                            .color(Color::Muted),
-                                    )
-                                    .child(
-                                        Label::new("/").size(LabelSize::Small).color(
-                                            Color::Custom(
-                                                cx.theme().colors().text_muted.opacity(0.4),
-                                            ),
-                                        ),
-                                    )
-                                })
                                 .child(
-                                    Label::new(branch_name)
+                                    Icon::new(IconName::GitWorktree)
+                                        .size(IconSize::XSmall)
+                                        .color(Color::Muted),
+                                )
+                                .child(
+                                    Label::new(worktree_label)
                                         .size(LabelSize::Small)
                                         .color(Color::Muted),
                                 ),
                         ),
                     move |_window, cx| {
                         Tooltip::with_meta(
-                            "Git Switcher",
-                            Some(&zed_actions::git::Branch),
-                            "Worktrees, Branches, and Stashes",
+                            "Worktree",
+                            Some(&zed_actions::git::Worktree),
+                            format!("Currently In Use: {}", worktree_tooltip_label),
                             cx,
                         )
                     },
                 )
-                .anchor(gpui::Corner::TopLeft),
+                .anchor(gpui::Corner::TopLeft)
+        };
+
+        let branch_tooltip_label = branch_name.clone();
+
+        let git_picker_button = PopoverMenu::new("branch-menu")
+            .menu(move |window, cx| {
+                Some(git_ui::git_picker::popover(
+                    workspace.downgrade(),
+                    effective_repository.clone(),
+                    git_ui::git_picker::GitPickerTab::Branches,
+                    gpui::rems(34.),
+                    window,
+                    cx,
+                ))
+            })
+            .trigger_with_tooltip(
+                ButtonLike::new("project_branch_trigger")
+                    .selected_style(ButtonStyle::Tinted(TintColor::Accent))
+                    .child(
+                        h_flex()
+                            .gap_0p5()
+                            // todo dl: remove this setting?!
+                            // .when(settings.show_branch_icon, |this| {
+                            //     let (icon, icon_color) = icon_info;
+                            //     this.child(Icon::new(icon).size(IconSize::XSmall).color(icon_color))
+                            // })
+                            .child(
+                                Icon::new(IconName::GitBranch)
+                                    .size(IconSize::XSmall)
+                                    .color(Color::Muted),
+                            )
+                            .child(
+                                Label::new(branch_name)
+                                    .size(LabelSize::Small)
+                                    .color(Color::Muted),
+                            ),
+                    ),
+                move |_window, cx| {
+                    let meta = if is_detached_head {
+                        "In Detached HEAD State".to_string()
+                    } else {
+                        format!("Currently Checked Out: {}", branch_tooltip_label)
+                    };
+                    Tooltip::with_meta("Branch & Stash", Some(&zed_actions::git::Branch), meta, cx)
+                },
+            )
+            .anchor(gpui::Corner::TopLeft);
+
+        Some(
+            h_flex()
+                .gap_0p5()
+                .child(worktree_button)
+                .child(
+                    Label::new("/")
+                        .size(LabelSize::Small)
+                        .color(Color::Custom(cx.theme().colors().text_muted.opacity(0.4))),
+                )
+                .child(git_picker_button)
+                .into_any_element(),
         )
     }
